% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moex_api.R
\name{issBoards}
\alias{issBoards}
\title{Get boards list for a market}
\usage{
issBoards(engine = "stock", market = "shares", raw = F)
}
\arguments{
\item{raw}{}
}
\value{

}
\description{
Get boards list for a market
}
