% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moex_api.R
\name{issCandles}
\alias{issCandles}
\title{OHLC data for a security.}
\usage{
issCandles(
  engine = "stock",
  market = "shares",
  board = "TQBR",
  security = "VTBR",
  from = "2020-01-01",
  til = "2037-12-31",
  interval = 10
)
}
\arguments{
\item{interval}{}
}
\value{

}
\description{
OHLC data for a security.
}
