% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issGcurve.R
\name{issGcurve}
\alias{issGcurve}
\title{G-Curve parametres}
\usage{
issGcurve(date = NULL, intraday = F, time = "")
}
\arguments{
\item{date}{a date in the past}

\item{intraday}{if TRUE, the function returns all intraday coefficients}

\item{time}{the last known coefficients set for the given time}
}
\value{

}
\description{
G-Curve parametres
}
