% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moex_api.R
\name{issGetBonds}
\alias{issGetBonds}
\title{Bonds data}
\usage{
issGetBonds(board = "", sectypes = "", tablenames = "")
}
\arguments{
\item{sectypes}{futures ticker or underlying}

\item{tablenames}{securities, marketdata, marketdata_yields, dataversion are avaliable}

\item{boards}{specify trading board (TQOB, EQOB, TQCB, etc.)}
}
\value{

}
\description{
Bonds data
}
