% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moex_api.R
\name{issGetFutures}
\alias{issGetFutures}
\title{FORTS futures data}
\usage{
issGetFutures(sectypes = "", tablenames = "")
}
\arguments{
\item{sectypes}{futures ticker or underlying}

\item{tablenames}{}
}
\value{

}
\description{
FORTS futures data
}
