% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moex_api.R
\name{issGetOptions}
\alias{issGetOptions}
\title{FORTS options data}
\usage{
issGetOptions(sectypes = "", tablenames = "")
}
\arguments{
\item{sectypes}{futures ticker or underlying}

\item{tablenames}{}
}
\value{

}
\description{
FORTS options data
}
