% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{issHistorySecurities}
\alias{issHistorySecurities}
\title{Historical data for securities at the special date}
\usage{
issHistorySecurities(
  engine = "stock",
  market = "bonds",
  board = "",
  date = "2020-06-01"
)
}
\arguments{
\item{board}{specify trading board (TQOB, EQOB, TQCB, etc.)}

\item{date}{}
}
\value{

}
\description{
Historical data for securities at the special date
}
