% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{issHistoryYields}
\alias{issHistoryYields}
\title{Historical bond yields}
\usage{
issHistoryYields(board = "", date = "2020-06-01")
}
\arguments{
\item{board}{specify trading board (TQOB, EQOB, TQCB, etc.)}

\item{date}{}
}
\value{

}
\description{
Historical bond yields
}
