% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moex_api.R
\name{issMarkets}
\alias{issMarkets}
\title{Returns a list of the markets for a given engine}
\usage{
issMarkets(engine = "stock", raw = F)
}
\arguments{
\item{raw}{}
}
\value{

}
\description{
Returns a list of the markets for a given engine
}
