% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moex_api.R
\name{issReference}
\alias{issReference}
\title{List of reference pages}
\usage{
issReference()
}
\value{
Function returns the referce page, structured in a table.
}
\description{
List of reference pages
}
\examples{

}
