% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moex_api.R
\name{issSecurities}
\alias{issSecurities}
\title{Get a security list fot a market or a board}
\usage{
issSecurities(engine = "stock", market = "shares", board = "", raw = F)
}
\arguments{
\item{raw}{}
}
\value{

}
\description{
Get a security list fot a market or a board
}
