% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moex_api.R
\name{issTrades}
\alias{issTrades}
\title{Loads all trades for a given security / board / market}
\usage{
issTrades(
  engine = "stock",
  market = "shares",
  board = "EQBR",
  security = "VTBR",
  raw = F
)
}
\arguments{
\item{raw}{}
}
\value{

}
\description{
Loads all trades for a given security / board / market
}
