% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moex_api.R
\name{Load_All_Pages}
\alias{Load_All_Pages}
\title{Load all data for a multipager}
\usage{
Load_All_Pages(path, tablename)
}
\arguments{
\item{tablename}{}
}
\value{

}
\description{
Load all data for a multipager
}
