% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moex_api.R
\name{MakeTable}
\alias{MakeTable}
\title{Make tables from xml nodes <data>}
\usage{
MakeTable(iss_document_data)
}
\arguments{
\item{iss_document_data}{}
}
\value{

}
\description{
Make tables from xml nodes <data>
}
