% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moex_api.R
\name{Read_ISS}
\alias{Read_ISS}
\title{Basic funtion - returns data from any iss adress}
\usage{
Read_ISS(path = "", tablenames = "")
}
\arguments{
\item{tablenames}{}
}
\value{

}
\description{
Basic funtion - returns data from any iss adress
}
